+++ src/main/java/roomescape/service/theme/module/ThemeResisterService.java
@@ -1,0 +1,34 @@
+package roomescape.service.theme.module;
+
+import org.springframework.stereotype.Service;
+import roomescape.domain.theme.Theme;
+import roomescape.domain.theme.ThemeName;
+import roomescape.dto.theme.ThemeRequest;
+import roomescape.repository.ThemeRepository;
+
+@Service
+public class ThemeResisterService {
+
+    private final ThemeRepository themeRepository;
+
+    public ThemeResisterService(ThemeRepository themeRepository) {
+        this.themeRepository = themeRepository;
+    }
+
+    public Long resisterTheme(ThemeRequest themeRequest) {
+        ThemeName name = new ThemeName(themeRequest.name());
+        validateNameDuplicate(name);
+        Theme theme = themeRequest.toEntity();
+
+        return themeRepository.save(theme).getId();
+    }
+
+    public void validateNameDuplicate(ThemeName name) {
+        if (themeRepository.existsByThemeName(name)) {
+            throw new IllegalArgumentException(
+                    "[ERROR] 동일한 이름의 테마가 존재해 등록할 수 없습니다.",
+                    new Throwable("theme_name : " + name)
+            );
+        }
+    }
+}
